requirements.txt
numpy==1.23.2
Pillow==9.1.0
Pillow==10.3.0
